+++ tests/fixtures/optimize-package-imports/barrels/extensionless/index.js
@@ -1,0 +1,2 @@
+export { default as Alarm } from "./icons/Alarm";
+export * from "./more";